backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
ortools==9.8.3296
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10
xmltodict==13.0.0
pyjwt==2.8.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
